Update lab_4.qmd
Question 1 was emphasising the contrast in the behaviour of usage of trash vs deleting permanently in the Windows and terminal, which was inconsistent. 

Instead we should compare the behaviour seen when deleting files from the graphical file managers (moved to trash) and when deleting using the rm command in the terminal (deleted permanently).

It's not dependent on the OS but how you are deleting it (using gui mode or in terminal). Even if you delete using the rm command in windows in the powershell it will be removed from the system completely. So I have replaced this inconsistent comparison with a comparison between gui mode and deleting from the terminal.